Moved all global styling to one file
